Move the registration manager to the API
It can be useful when implementing new services
